файл зависимостей - requirements.txt
Flask==2.1.2
requests==2.27.1
python-dotenv==0.19.2
pip install -r requirements.txt